--- react-app/src/components/ConCurrRedeem/ConCurrRedeem.jsx
@@ -1,11 +1,0 @@
-// import contractCall from "../ContractCall/ContractCall";
-
-// import React from 'react'
-
-// function ConCurrRedeem() {
-//   return (
-//     <div>ConCurrRedeem</div>
-//   )
-// }
-
-// export default ConCurrRedeem
