--- react-app/src/components/Posts/AllPosts/Posts.css
@@ -1,27 +1,0 @@
-.all-posts-container {
-    display: flex;
-    flex-direction: column;
-}
-
-.post-detail-container {
-    margin: 10px 0;
-    padding: 10px;
-    border: 1px solid var(--ethblue);
-    border-radius: 5px;
-    background-color: rgba(220,220,220, 0.5);
-}
-
-#post-username {
-    padding-bottom: 10px;
-    font-size: 10px;
-}
-
-#post-title {
-    font-size: 16px;
-    color: var(--ethblue);
-}
-
-#post-caption {
-    font-size: 13px;
-    color: var(--ethlightblue);
-}
